requirements.txt
beautifulsoup4==4.12.2
lxml==5.3.0
python-dotenv==1.0.0
requests==2.31.0
requests-cache==1.2.1
python-docx==1.1.2
schedule==1.2.2